+++ rust/src/plugins/mod.rs
@@ -1,3 +1,5 @@
 pub mod evaluate_plugin;
 
 pub mod vector;
+
+pub mod assert_equal;