+++ rust/src/plugins/vector.rs
@@ -1,16 +1,32 @@
 use num_bigint::BigUint;
+use std::collections::HashMap;
 use std::convert::TryFrom;
 
 use crate::structs::count::Count;
-use crate::Result;
+use crate::{Result, TypeId};
 
+/// This function performs the following checks on vector_add/mul inputs.
+/// - there is no public/private inputs
+/// - `params` are compliant with the plugin vector and the operation add/mul
+/// - `output_count` and `input_count` are compliant with `plugin(vector, add/mul, params)`
+/// - `inputs` is compliant with `plugin(vector, add/mul, params)`
 fn vector_check<'a>(
     output_count: &'a [Count],
     input_count: &'a [Count],
     inputs: &'a [&BigUint],
+    public_inputs: &HashMap<TypeId, Vec<BigUint>>,
+    private_inputs: &HashMap<TypeId, Vec<BigUint>>,
     params: &'a [String],
     moduli: &'a [BigUint],
 ) -> Result<(usize, &'a BigUint)> {
+    // Check that there is no public/private inputs
+    if !public_inputs.is_empty() {
+        return Err("plugin(vector, add/mul) does not consume any public input.".into());
+    }
+    if !private_inputs.is_empty() {
+        return Err("plugin(vector, add/mul) does not consume any private input.".into());
+    }
+
     // Check that params are compliant with the plugin vector and the operation add/mul
     if params.len() != 2 {
         return Err(
@@ -19,7 +35,6 @@ fn vector_check<'a>(
     }
     let param_type_id = params[0].parse::<u8>()?;
     let param_len = params[1].parse::<usize>()?;
-    let param_len_u64 = u64::try_from(param_len)?;
     if param_len == 0 {
         return Err("plugin(vector, add/mul) cannot be called without inputs.".into());
     }
@@ -31,7 +46,7 @@ fn vector_check<'a>(
     })?;
 
     // Check that `output_count` and `input_count` are compliant with `plugin(vector, add/mul, params)`
-    let expected_output_count = vec![Count::new(param_type_id, param_len_u64)];
+    let expected_output_count = vec![Count::new(param_type_id, u64::try_from(param_len)?)];
     if *output_count != expected_output_count {
         return Err(format!(
             "When calling the plugin(vector, add/mul, {}, {}), the out parameter in the function signature must be equal to {:?} (and not {:?}).",
@@ -41,13 +56,13 @@ fn vector_check<'a>(
     }
 
     let expected_input_count = vec![
-        Count::new(param_type_id, param_len_u64),
-        Count::new(param_type_id, param_len_u64),
+        Count::new(param_type_id, u64::try_from(param_len)?),
+        Count::new(param_type_id, u64::try_from(param_len)?),
     ];
     if *input_count != expected_input_count {
         return Err(format!(
             "When calling the plugin(vector, add/mul, {}, {}), the in parameter in the function signature must be equal to {:?} (and not {:?}).",
-            param_type_id, param_len_u64, expected_input_count, input_count
+            param_type_id, param_len, expected_input_count, input_count
         )
             .into());
     }
@@ -70,10 +85,20 @@ pub fn vector_add(
     output_count: &[Count],
     input_count: &[Count],
     inputs: &[&BigUint],
+    public_inputs: &HashMap<TypeId, Vec<BigUint>>,
+    private_inputs: &HashMap<TypeId, Vec<BigUint>>,
     params: &[String],
     moduli: &[BigUint],
 ) -> Result<Vec<BigUint>> {
-    let (param_len, modulo) = vector_check(output_count, input_count, inputs, params, moduli)?;
+    let (param_len, modulo) = vector_check(
+        output_count,
+        input_count,
+        inputs,
+        public_inputs,
+        private_inputs,
+        params,
+        moduli,
+    )?;
 
     // Evaluate plugin(vector, add)
     let mut result = vec![];
@@ -90,10 +115,20 @@ pub fn vector_mul(
     output_count: &[Count],
     input_count: &[Count],
     inputs: &[&BigUint],
+    public_inputs: &HashMap<TypeId, Vec<BigUint>>,
+    private_inputs: &HashMap<TypeId, Vec<BigUint>>,
     params: &[String],
     moduli: &[BigUint],
 ) -> Result<Vec<BigUint>> {
-    let (param_len, modulo) = vector_check(output_count, input_count, inputs, params, moduli)?;
+    let (param_len, modulo) = vector_check(
+        output_count,
+        input_count,
+        inputs,
+        public_inputs,
+        private_inputs,
+        params,
+        moduli,
+    )?;
 
     // Evaluate plugin(vector, mul)
     let mut result = vec![];
@@ -115,7 +150,16 @@ fn test_vector_check() {
     ];
     let moduli = [BigUint::from_bytes_le(&[7])];
     let params = ["0".to_string(), "2".to_string()];
-    let result = vector_check(&output_count, &input_count, &inputs, &params, &moduli).unwrap();
+    let result = vector_check(
+        &output_count,
+        &input_count,
+        &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
+        &params,
+        &moduli,
+    )
+    .unwrap();
     let expected_result = (2_usize, &BigUint::from_bytes_le(&[7]));
     assert_eq!(result, expected_result);
 
@@ -125,6 +169,8 @@ fn test_vector_check() {
         &output_count,
         &input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &incorrect_params,
         &moduli,
     );
@@ -136,6 +182,8 @@ fn test_vector_check() {
         &output_count,
         &input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &incorrect_params,
         &moduli,
     );
@@ -147,6 +195,8 @@ fn test_vector_check() {
         &output_count,
         &input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &incorrect_params,
         &moduli,
     );
@@ -158,6 +208,8 @@ fn test_vector_check() {
         &incorrect_output_count,
         &input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &params,
         &moduli,
     );
@@ -169,6 +221,8 @@ fn test_vector_check() {
         &output_count,
         &incorrect_input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &params,
         &moduli,
     );
@@ -184,6 +238,8 @@ fn test_vector_check() {
         &output_count,
         &input_count,
         &incorrect_inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &params,
         &moduli,
     );
@@ -202,7 +258,16 @@ fn test_vector_add() {
     ];
     let moduli = [BigUint::from_bytes_le(&[7])];
     let params = ["0".to_string(), "2".to_string()];
-    let result = vector_add(&output_count, &input_count, &inputs, &params, &moduli).unwrap();
+    let result = vector_add(
+        &output_count,
+        &input_count,
+        &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
+        &params,
+        &moduli,
+    )
+    .unwrap();
     let expected_result = vec![BigUint::from_bytes_le(&[4]), BigUint::from_bytes_le(&[6])];
     assert_eq!(result, expected_result);
 
@@ -212,6 +277,8 @@ fn test_vector_add() {
         &output_count,
         &input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &incorrect_params,
         &moduli,
     );
@@ -230,7 +297,16 @@ fn test_vector_mul() {
     ];
     let moduli = [BigUint::from_bytes_le(&[7])];
     let params = ["0".to_string(), "2".to_string()];
-    let result = vector_mul(&output_count, &input_count, &inputs, &params, &moduli).unwrap();
+    let result = vector_mul(
+        &output_count,
+        &input_count,
+        &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
+        &params,
+        &moduli,
+    )
+    .unwrap();
     let expected_result = vec![BigUint::from_bytes_le(&[3]), BigUint::from_bytes_le(&[1])];
     assert_eq!(result, expected_result);
 
@@ -240,6 +316,8 @@ fn test_vector_mul() {
         &output_count,
         &input_count,
         &inputs,
+        &HashMap::new(),
+        &HashMap::new(),
         &incorrect_params,
         &moduli,
     );